NLP_task_7_Lexical_Similarity
Task 7. First, low and high frequency words (for threshold = 5) should be determined in a Turkish collection. Words with low frequency will be assumed to have spelling errors. To solve this problem, a Lexical Similarity function will be used (student can find at internet or prepare) and it will be suggested that these erroneous assumed words can change to the most lexically similar word from high-frequency words. The code to be prepared must give a two-column list: low-frequency words in the first column and the most lexically similar high-frequency words in the second column.